codeforces_tool_app_server_bk/server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose[cryptography]
cloudscraper
beautifulsoup4
requests
numpy